+++ src/client/resources/debugify.client.mixins.json
@@ -41,8 +41,6 @@
     "gameplay.mc159163.ClientPacketListenerMixin"
   ],
   "mixins": [
-    "basic.mc170462.MobEffectsMixin",
-    "basic.mc176559.ItemStackAccessor",
-    "basic.mc84661.MobEffectsMixin"
+    "basic.mc176559.ItemStackAccessor"
   ]
 }
