+++ src/main/resources/debugify.mixins.json
@@ -18,6 +18,7 @@
     "basic.mc132878.ArmorStandMixin",
     "basic.mc155509.PufferfishMixin",
     "basic.mc160095.CactusBlockMixin",
+    "basic.mc170462.MobEffectsMixin",
     "basic.mc179072.SwellGoalMixin",
     "basic.mc183990.MobMixin",
     "basic.mc199467.MthMixin",
@@ -33,6 +34,7 @@
     "basic.mc30391.LivingEntityMixin",
     "basic.mc69216.ServerPlayerMixin",
     "basic.mc7569.RconConsoleSourceMixin",
+    "basic.mc84661.MobEffectsMixin",
     "basic.mc88371.DragonLandingPhaseMixin",
     "basic.mc89146.ChunkAccessMixin",
     "basic.mc93018.AnimalMixin",
